Update: parse TodoList to a single component
